+++ test/cram/generate-projects.t/run.t
@@ -1,4 +1,4 @@
-  $ okra generate --conf=conf.yaml --no-activity | tail -n +3 | head -n 6
+  $ okra generate --conf=conf.yaml --no-activity --print-projects | tail -n +3 | head -n 6
   # Projects
   
   - Project 1 (#1)