+++ CHANGES.md
@@ -11,6 +11,8 @@
 - Improve the "Invalid time" error messages (#199, @gpetiot)
 - okra team lint: only print details of invalid/missing files, or total of valid files (#200, @gpetiot)
 - okra gen report: PR/issue entries formatted the same way as in engineer reports (#201, @gpetiot)
+- List of projects not printed by default anymore in generated reports (#212, @gpetiot).
+  Use the new option `--print-projects` to display the list.
 
 ### Fixed